Check if libfab file exists and then update lnet

+++ scripts/env/dev/init.sh
@@ -257,12 +257,15 @@ then
 fi
 
 #Fetch eth interface value form /etc/libfab.conf, add it to lnet and restart lnet
-iface=$(cat /etc/libfab.conf | cut -d "(" -f2 | cut -d ")" -f1)
-echo "options lnet networks=tcp($iface) config_on_load=1" > '/etc/modprobe.d/lnet.conf'
-service lnet restart
-if ! lctl list_nids ; then
-  echo "lctl list_nids returned empty"
-  exit 1
+libfab='/etc/libfab.conf'
+if [ -f $libfab ] ; then
+  iface=$(cat /etc/libfab.conf | cut -d "(" -f2 | cut -d ")" -f1)
+  echo "options lnet networks=tcp($iface) config_on_load=1" > '/etc/modprobe.d/lnet.conf'
+  service lnet restart
+  if ! lctl list_nids ; then
+    echo "lctl list_nids returned empty"
+    exit 1
+  fi
 fi
 
 # install all rpms which requires gcc as dependency
